Added a SORT OF higher order component that is currently rendering myactivities
